ARC-707: extract push error handlers into a common file
