+++ src/sqs/queues.ts
@@ -1,7 +1,8 @@
 import envVars from "../config/env";
 import {SqsQueue} from "./index";
 import {BackfillMessagePayload, backfillQueueMessageHandlerFactory} from "./backfill";
-import {PushQueueMessagePayload, pushQueueMessageHandler, pushQueueErrorHandler} from "./push";
+import {PushQueueMessagePayload, pushQueueMessageHandler} from "./push";
+import {jiraOctokitErrorHandler} from './error-handlers';
 import backfillQueueSupplier from "../backfill-queue-supplier";
 
 const LONG_POLLING_INTERVAL_SEC = 3;
@@ -15,7 +16,7 @@ const sqsQueues = {
 		maxAttempts: 3
 	},
 	backfillQueueMessageHandlerFactory(() => backfillQueueSupplier.supply()),
-	pushQueueErrorHandler
+	jiraOctokitErrorHandler
 	),
 
 	push: new SqsQueue<PushQueueMessagePayload>({ queueName: "push",
@@ -23,7 +24,7 @@ const sqsQueues = {
 		queueRegion: envVars.SQS_PUSH_QUEUE_REGION,
 		longPollingIntervalSec: LONG_POLLING_INTERVAL_SEC,
 		timeoutSec: 60,
-		maxAttempts: 5}, pushQueueMessageHandler,	pushQueueErrorHandler),
+		maxAttempts: 5}, pushQueueMessageHandler,	jiraOctokitErrorHandler),
 
 
 	start: () => {
