+++ test/unit/sqs/error-handlers.test.ts
@@ -1,5 +1,5 @@
 import statsd from "../../../src/config/statsd";
-import {pushQueueErrorHandler, PushQueueMessagePayload} from "../../../src/sqs/push";
+import {jiraOctokitErrorHandler} from "../../../src/sqs/error-handlers";
 import {Context} from "../../../src/sqs/index";
 import {getLogger} from "../../../src/config/logger";
 import {JiraClientError} from "../../../src/jira/client/axios";
@@ -7,7 +7,7 @@ import {RateLimitingError} from "../../../src/config/enhance-octokit";
 import {Octokit} from "probot";
 
 
-describe("Push Queue Error Handler", () => {
+describe("jiraOktokitErrorHandler", () => {
 
 	let statsdIncrementSpy = jest.spyOn(statsd, "histogram");
 
@@ -37,7 +37,7 @@ describe("Push Queue Error Handler", () => {
 		webhookId: "string"
 	};
 
-	const createContext = (receiveCount: number, lastAttempt: boolean): Context<PushQueueMessagePayload> =>
+	const createContext = (receiveCount: number, lastAttempt: boolean): Context<any> =>
 		({
 			receiveCount, lastAttempt, log: getLogger("test"), message: {}, payload: mockPayload
 		})
@@ -45,7 +45,7 @@ describe("Push Queue Error Handler", () => {
 
 	it("Returns normal retry when error is unknown", async () => {
 
-		const result = await pushQueueErrorHandler(new Error(), createContext(1, false));
+		const result = await jiraOctokitErrorHandler(new Error(), createContext(1, false));
 
 		expect(result.retryable).toBe(true)
 		expect(result.retryDelaySec).toBe(3*60)
@@ -54,7 +54,7 @@ describe("Push Queue Error Handler", () => {
 
 	it("Exponential backoff works", async () => {
 
-		const result = await pushQueueErrorHandler(new Error(), createContext(3, false));
+		const result = await jiraOctokitErrorHandler(new Error(), createContext(3, false));
 
 		expect(result.retryable).toBe(true)
 		expect(result.retryDelaySec).toBe(27*60)
@@ -63,7 +63,7 @@ describe("Push Queue Error Handler", () => {
 
 	it("Sends metrics when it was last attempt", async () => {
 
-		await pushQueueErrorHandler(new Error(), createContext(1, true));
+		await jiraOctokitErrorHandler(new Error(), createContext(1, true));
 
 		expect(statsdIncrementSpy).toBeCalledTimes(1);
 	});
@@ -78,34 +78,34 @@ describe("Push Queue Error Handler", () => {
 
 	it("Unretryable and no failure metric on Jira 401", async () => {
 
-		const result = await pushQueueErrorHandler(getJiraClientError(401), createContext(1, true));
+		const result = await jiraOctokitErrorHandler(getJiraClientError(401), createContext(1, true));
 		expect(result.retryable).toBe(false)
 		expect(statsdIncrementSpy).toBeCalledTimes(0);
 	});
 
 	it("Unretryable and no failure metric on Jira 403", async () => {
 
-		const result = await pushQueueErrorHandler(getJiraClientError(403), createContext(1, true));
+		const result = await jiraOctokitErrorHandler(getJiraClientError(403), createContext(1, true));
 		expect(result.retryable).toBe(false)
 		expect(statsdIncrementSpy).toBeCalledTimes(0);
 	});
 
 	it("Unretryable and no failure metric on Jira 404", async () => {
 
-		const result = await pushQueueErrorHandler(getJiraClientError(404), createContext(1, true));
+		const result = await jiraOctokitErrorHandler(getJiraClientError(404), createContext(1, true));
 		expect(result.retryable).toBe(false)
 		expect(statsdIncrementSpy).toBeCalledTimes(0);
 	});
 
 	it("Retryable and no failure metric sent on Jira 500", async () => {
 
-		const result = await pushQueueErrorHandler(getJiraClientError(500), createContext(1, true));
+		const result = await jiraOctokitErrorHandler(getJiraClientError(500), createContext(1, true));
 		expect(result.retryable).toBe(true)
 		expect(statsdIncrementSpy).toBeCalledTimes(1);
 	});
 
 	it("Retryable with proper delay on Rate Limiting", async () => {
-		const result = await pushQueueErrorHandler(new RateLimitingError(Math.floor(new Date("2020-01-01").getTime()/1000) + 100), createContext(1, false));
+		const result = await jiraOctokitErrorHandler(new RateLimitingError(Math.floor(new Date("2020-01-01").getTime()/1000) + 100), createContext(1, false));
 		expect(result.retryable).toBe(true)
 		//Make sure delay is equal to recommended delay + 10 seconds
 		expect(result.retryDelaySec).toBe(110)
@@ -116,7 +116,7 @@ describe("Push Queue Error Handler", () => {
 
 		const error : Octokit.HookError = {...new Error("Err"), status: 401, headers: {}}
 
-		const result = await pushQueueErrorHandler(error, createContext(1, true));
+		const result = await jiraOctokitErrorHandler(error, createContext(1, true));
 		expect(result.retryable).toBe(false)
 		expect(statsdIncrementSpy).toBeCalledTimes(0);
 	});
@@ -125,7 +125,7 @@ describe("Push Queue Error Handler", () => {
 
 		const error : Octokit.HookError = {...new Error("Err"), status: 500, headers: {}}
 
-		const result = await pushQueueErrorHandler(error, createContext(1, true));
+		const result = await jiraOctokitErrorHandler(error, createContext(1, true));
 		expect(result.retryable).toBe(true)
 		expect(statsdIncrementSpy).toBeCalledTimes(1);
 	});
